+++ frontend/models/ContactForm.php
@@ -24,10 +24,13 @@ class ContactForm extends Model
         return [
             // name, email, subject and body are required
             [['name', 'email', 'subject', 'body', 'verifyCode'], 'required'],
+            // We need to sanitize them
+            [['name', 'subject', 'body'], 'filter', 'filter' => 'strip_tags'],
             // email has to be a valid email address
             ['email', 'email'],
             // verifyCode needs to be entered correctly
-            ['verifyCode', 'captcha']
+            ['verifyCode', 'captcha'],
+
         ];
     }