+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Yii Starter Kit Change Log
 
 dev-master
 ----------
+- Added some sanitizing in ContactForm (#339)
 - env() helper function added
 - trntv/yii2-glide and trntv/yii2-command-bus versions updated
 - Testing improvements